refactor auth route into controller
